+++ src/ex/ex-stream.cpp
@@ -63,14 +63,10 @@
 
 #include <regex>
 
-const int default_line_size = 1000;
-
 wex::ex_stream::ex_stream(wex::ex* ex)
   : m_context_lines(50)
   , m_buffer_size(1000000)
   , m_buffer(new char[m_buffer_size])
-  , m_current_line_size(default_line_size)
-  , m_current_line(new char[m_current_line_size])
   , m_ex(ex)
   , m_stc(ex->get_stc())
 {
@@ -142,9 +138,8 @@ bool wex::ex_stream::erase(const addressrange& range)
 
 void wex::ex_stream::filter_line(int start, int end, std::streampos spos)
 {
-  // m_buffer , now from end of m_buffer search
-  // backwards for newline, this is the m_current_line to handle, and set the
-  // stream pointer to position before that newline.
+  // copy from start of m_buffer to the current line
+  // and set the stream pointer
   const size_t sz(end - start);
 
   strncpy(m_current_line, m_buffer + start + 1, sz);
@@ -153,7 +148,7 @@ void wex::ex_stream::filter_line(int start, int end, std::streampos spos)
 
   if (spos == 0)
   {
-    m_current_line_size = start - 1;
+    m_current_line_size = start;
     m_stream->clear();
     m_stream->seekg(0);
   }
@@ -218,6 +213,8 @@ bool wex::ex_stream::find_data(data::find& f)
 
   m_stream->clear();
 
+  m_pos_to_bol = true;
+
   // Notice we start get..line, and not searching in the current line.
   while (!found && ((f.is_forward() && get_next_line()) ||
                     (!f.is_forward() && get_previous_line())))
@@ -227,6 +224,10 @@ bool wex::ex_stream::find_data(data::find& f)
       (use_regex && std::regex_search(m_current_line, r)))
     {
       found = true;
+    }
+    else
+    {
+      m_pos_to_bol = false;
     }
   }
 
@@ -274,7 +275,7 @@ bool wex::ex_stream::find_finish(data::find& f, bool& found)
     log::trace("ex stream found") << f.text() << "current" << m_line_no;
   }
 
-  m_current_line_size = default_line_size;
+  m_current_line_size = m_default_line_size;
 
   return found;
 }
@@ -325,7 +326,7 @@ int wex::ex_stream::get_line_count_request()
 
     if (m_block_mode)
     {
-      line_no += m_buffer_size / default_line_size;
+      line_no += m_buffer_size / m_default_line_size;
     }
   }
 
@@ -378,24 +379,35 @@ bool wex::ex_stream::get_previous_line()
     m_stream->seekg((size_t)pos - m_current_line_size);
     pos = m_stream->tellg();
   }
-  else
+  else if (pos > 0)
   {
-    pos = 0;
     m_stream->seekg(0);
+    m_current_line_size = pos;
+    m_current_line_size--;
+    pos = 0;
   }
 
   m_stream->read(m_buffer, m_current_line_size);
 
   if (m_stream->gcount() > 0)
   {
-    const int end = m_stream->gcount() - 1;
+    int  end    = m_stream->gcount() - 1;
+    bool second = !m_pos_to_bol;
 
     for (int i = end; i >= 0; i--)
     {
       if (m_buffer[i] == '\n')
       {
-        filter_line(i, end, pos);
-        return true;
+        if (!second)
+        {
+          end    = i;
+          second = true;
+        }
+        else
+        {
+          filter_line(i, end, pos);
+          return true;
+        }
       }
     }
 
@@ -410,7 +422,7 @@ bool wex::ex_stream::get_previous_line()
     }
 
     // There was no newline, this implies block mode.
-    if (m_current_line_size == default_line_size)
+    if (m_current_line_size == m_default_line_size)
     {
       m_block_mode = true;
       return static_cast<int>(m_stream->gcount()) > m_current_line_size - 1;
@@ -600,13 +612,17 @@ void wex::ex_stream::set_text()
     m_stc->GetLineEndPosition(m_stc->GetLineCount() - 1));
 }
 
-void wex::ex_stream::stream(file& f)
+void wex::ex_stream::stream(file& f, size_t default_line_size)
 {
   if (!f.is_open())
   {
     log("file is not open") << f.path();
     return;
   }
+
+  m_default_line_size = default_line_size;
+  m_current_line_size = default_line_size;
+  m_current_line      = new char[m_current_line_size];
 
   m_file   = &f;
   m_stream = &f.stream();